backend_flask/requirements.txt
Flask
google-generativeai
langchain
langchain-google-genai
langchain-community
pypdf
chromadb
Flask-PyMongo
Flask-JWT-Extended
bcrypt
pytube
openai